==> s6_fields_of_study.ipynb <==
!pip install cairocffi
import cairocffi
import pandas as pd
import dask
from dask import dataframe as dd
import igraph as ig
import numpy as np

header = '/mnt/e/MAG/mag-2021-01-05/advanced/'
fields_affil = 'data_temp/paper_fields_of_study_byPaperID.txt'
fields_schema = 'FieldOfStudyChildren.txt'
fields_infos = 'FieldsOfStudy.txt'

fields_affil_dd = dd.read_csv(fields_affil, header=None, sep='\t', names=['paper_id', 'field_id', 'score'])
print(fields_affil_dd.head())

fields_schema_dd = pd.read_csv(header+fields_schema, header=None, sep='\t', names=['field_id', 'field_child_id'])
print(fields_schema_dd.head())

pairs = []
for idx,row in fields_schema_dd.iterrows():
    try:
        pairs.append((str(row['field_child_id']), str(row['field_id'])))
    except:
        print(row)
        
fields_infos_dd = pd.read_csv(header+fields_infos, header=None, sep='\t')[[0, 1, 2]]
fields_infos_dd.columns = ['field_id', 'rank', 'normalized_name']
print(fields_infos_dd.head())

len(fields_affil_dd)

nodes = set()
for a,b in pairs:
    nodes.add(a)
    nodes.add(b)
    
fields_tree = ig.Graph(directed=True)
fields_tree.add_vertices(len(nodes))
fields_tree.vs['name'] = list(nodes)
fields_tree.add_edges(pairs)

idxs_out0 = [i for i, d in enumerate(fields_tree.degree(mode=ig.OUT)) if d == 0]
print(len(idxs_out0))
parents = [int(fields_tree.vs[i]['name']) for i in idxs_out0]
print(idxs_out0[:10])

# print parents ids
fields_infos_dd[fields_infos_dd['field_id'].isin(parents)].head(15)

def path_to_parent(g, v):
    idxs = g.neighbors(v, mode=ig.OUT)
    if len(idxs) == 0:
        return [g.vs[v]['name']]
    else:
        temp = []
        for i in idxs:
            temp += path_to_parent(g, i)
        return temp

def get_parents(g, v):
    temp = path_to_parent(g, v)
    parents_str = ','.join(temp)
    parents = [int(x) for x in temp]
    return parents, parents_str

# o parent é parent dele mesmo
print(fields_tree.vs[1077])
print(get_parents(fields_tree, 1077))

# output_file = open('data/fields_of_study.csv', 'w')
# for child in range(fields_tree.vcount()):
#     parents_ids, parents_id_str = get_parents(fields_tree, child)
#     parents_names = fields_infos_dd[fields_infos_dd[0].isin(parents_ids)][2].values
#     child_mag_id = fields_tree.vs[child]['name']
#     child_name = fields_infos_dd[fields_infos_dd[0] == int(child_mag_id)][2].values
#     output_file.write('%s\t%s\t%s\t%s\n' % (child_name[0], child_mag_id, ','.join(parents_names), parents_id_str))
    
# output_file.close()

# áreas que não estão na hierarquia
a = set([int(t) for t in fields_tree.vs['name']]) # quem está na hierarquia
b = set(fields_infos_dd['field_id'].values) # todas as áreas listadas no MAG
c = b-a
fields_infos_dd[fields_infos_dd['field_id'].isin(c)].head(10)

fields_of_study = pd.read_csv('data/fields_of_study.csv', sep='\t', header=None, names=['field', 'field_id', 'parents', 'parents_id'])
fields_of_study.head()

fields_affil_dd = fields_affil_dd.merge(fields_of_study, how='left', on='field_id')
fields_affil_dd.head()

# fields_affil_dd.to_csv('data_temp/FOS_split/fields_papers_*.csv', header=None)

import json
def get_fos(row):
    if row['weights'] != None:
        w = json.loads(row['weights'])
        if len(w) <= 0:
            return ''
        max_key = max(w, key=lambda k: w[k])
        
        return max_key
    return ''

    
# 1960 1970, 1980, 1990
for year in [1960]:
    authors_fos = dd.read_csv('data/AuthorsFOS_split/authors_weights_complete_year_%d*' % year, sep='\t', header=None, names=['author_id', 'weights'])
    valid_authors = pd.read_csv('data/valid_authors_full.txt', header=None, names=['author_id', 'papers', 'cits'])
    valid_authors = valid_authors['author_id'].apply(int)
    valid_authors = valid_authors.values
    print(valid_authors[:10])

    authors_fos = authors_fos[authors_fos['author_id'].isin(valid_authors)]
    print(authors_fos.head())
    authors_fos['fos'] = authors_fos.apply(get_fos, axis=1, meta=('str'))
    print(authors_fos.head())
    print(len(authors_fos), len(authors_fos['author_id'].unique()))
#     from dask.diagnostics import ProgressBar
#     with ProgressBar():
#         authors_fos.to_csv('data/valid_authors_%d_fos_10a_full.csv' % year, sep='\t', header=None, index=None, single_file=True)

import glob
for year in [2020]:
    print(glob.glob('data/AuthorsFOS_split/authors_fos_weights_%d_final_*' % year)[:10])
    
    authors_fos = dd.read_csv('data/AuthorsFOS_split/authors_fos_weights_%d_final_*' % year, sep='\t', header=None, names=['author_id', 'weights'])
    valid_authors = pd.read_csv('data/valid_authors_full.txt', header=None, names=['author_id', 'papers', 'cits'])
    valid_authors = valid_authors['author_id'].apply(int)
    valid_authors = valid_authors.values
    print(valid_authors[:10])

    authors_fos = authors_fos[authors_fos['author_id'].isin(valid_authors)]
    print(authors_fos.head())
    authors_fos['fos'] = authors_fos.apply(get_fos, axis=1, meta=('str'))
    print(authors_fos.head())
    print(len(authors_fos), len(authors_fos['author_id'].unique()))

    from dask.diagnostics import ProgressBar
    with ProgressBar():
        authors_fos.to_csv('data/valid_authors_%d_fos_10a_full.csv' % year, sep='\t', header=None, index=None, single_file=True)

import tqdm
fos = authors_fos_250k['fos']
X = []
for row in tqdm.tqdm(fos, total=len(valid_authors)):
    X.append(row)

import numpy as np
unique, count = np.unique(X, return_counts=True)
for u,c in zip(unique, count):
    print(u, c)

unique = [int(u) for u in unique]
print(fields_infos_dd.head())
unique_freq = fields_infos_dd[fields_infos_dd['field_id'].isin(unique)]
unique_freq['count'] = unique_freq.apply(lambda row: count[ unique.index(row['field_id']) ], axis=1)
print(unique_freq)

fields_infos_dd = pd.read_csv(header+fields_infos, header=None, sep='\t')[[0, 1, 2]]
fields_infos_dd.to_csv('data/fields_names.csv', sep='\t', header=None)
